% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_countries.R
\name{detect_countries}
\alias{detect_countries}
\title{Convert messy text and email data into standardized country names or country codes}
\usage{
detect_countries(data, id, text, output, email)
}
\arguments{
\item{data}{A data frame or data frame extension (e.g. a tibble).}

\item{id}{A numeric or character vector unique to each entry.}

\item{text}{Character vector of text data that includes the name of cities, states, 
and/or countries that will be standardized into country names or country codes.}

\item{output}{Country name or country code. Either "country_name", "iso2", or "iso3". 
If multiple countries are detected, they will be separated by the "|" symbol.}

\item{email}{Character vector of email or email domain information.}
}
\description{
This function standardizes messy text data that contains city, state, and/or country names 
as well as email domains into standardized country names. The detect_countries() function 
integrates tidyorgs's text_to_countries() and email_to_countries() together, providing 
the capacity to match users to countries for organizational or geographic analysis in a tidy framework.
}
\examples{

library(tidyverse)
library(tidyorgs)
data(github_users)

classified_users <- github_users \%>\%
  detect_countries(login, location, country_name, email)
  
}
