% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_nonprofit.R
\name{detect_nonprofit}
\alias{detect_nonprofit}
\title{Match messy text and email data to nonprofit organizations}
\usage{
detect_nonprofit(
  data,
  id,
  input,
  output,
  email = FALSE,
  country = FALSE,
  parent_org = FALSE,
  org_type = FALSE
)
}
\arguments{
\item{data}{A data frame or data frame extension (e.g. a tibble).}

\item{id}{A numeric or character vector unique to each entry.}

\item{input}{Character vector of messy or unstructured text that will be matched to nonprofit organizations.}

\item{output}{Output column to be created as string or symbol.}

\item{email}{Optional character vector of email or email domain information. Defaults to FALSE.}

\item{country}{Optional parameter that returns country of organization when available. Defaults to FALSE.}

\item{parent_org}{Optional parameter that returns the parent organization when available. 
For the nonprofit sector, this value defaults to FALSE.}

\item{org_type}{Optional parameter that returns organization type when available.}
}
\description{
This function standardizes messy text data and/or email information to nonprofit organizations. 
The detect_nonprofit() function iterates through email domains and unstructured text to match patterns in our 
curated dictionaries to standardize nonprofit organizations. This tool is designed to optimize pattern detection 
for in the linkage of multiple datasets, for bibliometric analysis, and for sector classification in 
social, economic, and policy analysis.
}
\examples{

library(tidyverse)
library(tidyorgs)
data(github_users)

classified_users <- github_users \%>\%
  detect_nonprofit(login, company, organization, email, parent_org, org_type)

}
