% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_orgs.R
\name{detect_orgs}
\alias{detect_orgs}
\title{Match messy text and email data to academic, business, government, or nonprofit organizations}
\usage{
detect_orgs(data, id, text, output, sector, email)
}
\arguments{
\item{data}{A data frame.}

\item{id}{A numeric or character vector unique to each entry.}

\item{text}{Character vector of messy or unstructured text that will be matched to organizations 
from one (or all) of four economic sectors (see sector parameter).}

\item{output}{Output column to be created as string or symbol.}

\item{sector}{Sector to match organizations. Either "all", "academic", "business", "government", 
or "nonprofit". Defaults to "all".}

\item{email}{Character vector of email or email domain information.}
}
\description{
This function standardizes messy text data and/or email information to social organizations. 
The detect_orgs() function integrates text_to_orgs() and email_to_orgs() together, providing 
the capacity to match users to organizations for social, economic, or policy analysis in a tidy framework.
}
\examples{

library(tidyverse)
library(tidyorgs)
data(github_users)

classified_users <- github_users \%>\%
  detect_orgs(login, company, organization, academic, email)

}
