% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_to_orgs.R
\name{email_to_orgs}
\alias{email_to_orgs}
\title{Match emails to social organizations}
\usage{
email_to_orgs(data, id, input, output, sector)
}
\arguments{
\item{data}{A data frame.}

\item{id}{A numeric or character vector unique to each entry.}

\item{input}{Character vector of emails or email domains.}

\item{output}{Desired name of classified organization column.}

\item{sector}{Choose "all", "academic", "business", "goverment", or "nonprofit". Defaults to "academic".}
}
\description{
Full explanation here...
}
\examples{

library(tidyverse)
library(tidyorgs)
data(github_users)

classified_by_email <- github_users \%>\%
  email_to_orgs(login, email, organization, academic)

}
