% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_orgs.R
\name{standardize_orgs}
\alias{standardize_orgs}
\title{Standardize organizations names across economic sectors}
\usage{
standardize_orgs(df, input, output, sector)
}
\arguments{
\item{input}{Character vector of text that contains unstandardized organization information}

\item{output}{Desired name of organization column.}

\item{sector}{Choose "academic", "all", "business", "goverment", or "nonprofit".}

\item{data}{A data frame.}
}
\description{
Full explanation here...
}
\examples{

library(tidyverse)
library(tidyorgs)
data(github_users)

classified <- github_users \%>\%
  standardize_orgs(login, company, academic)


}
