% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_to_countries.R
\name{text_to_countries}
\alias{text_to_countries}
\title{Match messy text data to countries using a 'funneling' method}
\usage{
text_to_countries(data, id, input, output)
}
\arguments{
\item{data}{A data frame or data frame extension (e.g. a tibble).}

\item{id}{A numeric or character vector unique to each entry.}

\item{input}{Character vector of messy or unstructured text that will
be unnested as n-grams and matched to dictionary of organizations in specified sector.}

\item{output}{Output column to be created as string or symbol.}
}
\description{
This function allocates entries of unstructured text data to countries 
extracting and iterating through consecuetive word sequences (or n-grams). 
To do this, the function extracts n-grams using the tidytext 
package, matching all sequences in the unstructured data that have n words 
and then 'funneling' through all sequences of n-1, n-2, etc. words before 
matching the single tokens. This function returns a dataframe akin to the 
original dataframe with one additional column aligning with the "output" parameter.
}
\examples{

library(tidyverse)
library(tidyorgs)
data(github_users)

classified_by_text <- github_users \%>\%
  text_to_countries(login, location, country_name)

}
