% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_to_orgs.R
\name{text_to_orgs}
\alias{text_to_orgs}
\title{Match messy text data to social organizations using a 'funneling' method}
\usage{
text_to_orgs(data, id, input, output, sector)
}
\arguments{
\item{data}{A data frame.}

\item{id}{A numeric or character vector unique to each entry.}

\item{input}{Character vector of messy or unstructed text that will
be unnested as n-grams and matched to dictionary of academic instiutions.}

\item{output}{Desired name of classified organization column.}

\item{sector}{Choose "all", "academic", "business", "government", or "nonprofit". Defaults to "academic".}
}
\description{
This function matches unstructured text data to a dictionary of 9,000+
academic, business, government, and nonprofit institutions by extracting
and iterating through consequetive word sequences (or n-grams). To do this,
the function extracts n-grams, matching all sequences in the unstructured
data that have 12 words (default) and then 'funneling' through all sequences
of 9, 8, etc. words before matching the single tokens. This process returns
a dataframe of ids, organizations, and sectors for only those rows matched
within the sector(s) specified.
}
\examples{

library(tidyverse)
library(tidyorgs)
data(github_users)

classified_by_text <- github_users \%>\%
  text_to_orgs(login, company, organization, academic)

}
