% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_to_sectors.R
\name{text_to_sectors}
\alias{text_to_sectors}
\title{Match messy text data to econonmic sectors}
\usage{
text_to_sectors(data, id, input, output, sector)
}
\arguments{
\item{data}{A data frame.}

\item{id}{A numeric or character vector unique to each entry.}

\item{input}{Character vector of messy or unstructed text.}

\item{output}{Desired name of classified organization column.}

\item{sector}{Choose "all", "academic", "business", "goverment", or "nonprofit". Defaults to "academic".}
}
\description{
Explanation here...
}
\examples{

library(tidyverse)
library(tidyorgs)
data(github_users)

classified_by_text <- github_users \%>\%
  text_to_sectors(login, company, organization, academic)

}
